Fix an multiplication overflow bug in fixed math.
When multiplying two numbers together that were both > 1, we'd overflow
the available 16 bits of integer available.

I'm not sure this is the best way around it, but for now I chop off 8
bits of fractional precision in the case that both numbers are > 1.
It's much less precise but better than totally wrong.